Exercise 4.21: Blog List Expansion, step 9

+++ part4/bloglist/controllers/blogs.js
@@ -41,6 +41,21 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
+  const blogToDelete = await Blog.findById(request.params.id);
+  if (!blogToDelete) {
+    return response.status(404).json({ error: "blog not found" });
+  }
+
+  // toString() of a Mongoose object is the object's ID
+  if (blogToDelete.user.toString() !== decodedToken.id) {
+    return response.status(401).json({ error: "invalid user" });
+  }
+
   await Blog.findByIdAndDelete(request.params.id);
   response.status(204).end();
 });
